src/test/java/com/github/agiledon/swift/collect/SwiftListTest.java
package com.github.agiledon.swift.collect;

import com.github.agiledon.swift.base.Predicates;
import org.junit.Test;

import java.util.List;

import static com.google.common.collect.Lists.newArrayList;
import static org.hamcrest.CoreMatchers.is;
import static org.junit.Assert.assertThat;

public class SwiftListTest {
    @Test
    public void should_split_to_three_partitions_by_given_predicates() {
        List<String> target = newArrayList(
                "first line",
                "second line",
                "///",
                "third line",
                "fourth line",
                "fifth line",
                "///",
                "sixth line",
                "seventh line"
        );

        List<List<String>> partitions = SwiftList.partition(target, new Predicates<String>() {
            @Override
            public boolean apply(String element) {
                return element.trim().startsWith("///");
            }
        });

        assertThat(partitions.size(), is(3));
        assertThat(partitions.get(0).get(0), is("first line"));
        assertThat(partitions.get(1).get(0), is("third line"));
        assertThat(partitions.get(2).get(0), is("sixth line"));
    }

}
